+++ app/Http/Controllers/MeasureController.php
@@ -51,8 +51,10 @@ class MeasureController extends Controller
     }
 
 
-    public function index()
+    public function index(Measure $measure)
     {
+        $this->authorize('view', $measure);
+
         $pagename = "دليل المقاسات";
         $formPage = "update-measures-form";
         $addNew = "تعديل الصورة";
@@ -74,14 +76,17 @@ class MeasureController extends Controller
 
     }
 
-    public function insertData()
+    public function insertData(Measure $measure)
     {
+        $this->authorize('view', $measure);
+
         $mediaLibrary = MediaLibrary::all();
         return view('update-measures-form')->with( compact('mediaLibrary'));
     }
 
-    public function fetch_image($id)
+    public function fetch_image($id,Measure $measure)
     {
+        $this->authorize('view', $measure);
 
         $image = Measure::findOrFail($id);
         $image_file = Image::make($image->medl_img_ved)->resize(60, 60);
@@ -90,8 +95,9 @@ class MeasureController extends Controller
         return $response;
     }
 
-    public function destroy($id)
+    public function destroy($id,Measure $measure)
     {
+        $this->authorize('view', $measure);
 
         $data = Measure::where("fakeId","=","$id")->first();;
         $data->delete();
@@ -99,8 +105,9 @@ class MeasureController extends Controller
         return redirect()->back();
     }
 
-    public function enableOrdisable($id)
+    public function enableOrdisable($id,Measure $measure)
     {
+        $this->authorize('view', $measure);
 
         $data = Measure::where("fakeId","=","$id")->first();;
 
@@ -117,8 +124,10 @@ class MeasureController extends Controller
     }
 
 
-    function store(Request $request)
+    function store(Request $request,Measure $measure)
     {
+        $this->authorize('view', $measure);
+
 //        $image = Image::make($request->file('medl_id'))->encode('jpeg');
 //        MeasureController::$storeImage = $image;
         $mesuImage = new MediaLibrary();
@@ -132,6 +141,8 @@ class MeasureController extends Controller
 
     public function update(Request $request, Measure $measure,$id)
     {
+        $this->authorize('view', $measure);
+
         $currentValues = measuresImages::where("fakeId","=","$id")->first();
         $mediaLibrary = MediaLibrary::all();
         $currentMedias = MediaLibrary::all()
@@ -143,7 +154,9 @@ class MeasureController extends Controller
     }
 
 
-    public function store_update(Request $request, $id){
+    public function store_update(Request $request, $id,Measure $measure){
+        $this->authorize('view', $measure);
+
         $data = Measure::where("fakeId","=","$id")->first();
         $data->update($request->all());
         return redirect('/measure');
@@ -162,7 +175,9 @@ class MeasureController extends Controller
 //   }
 
 
-    public function search(Request $request){
+    public function search(Request $request,Measure $measure){
+        $this->authorize('view', $measure);
+
         $key = trim($request->get('search'));
